ToolCalling_Agent/pyproject.toml
[project]
name = "toolcalling-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "black>=25.12.0",
    "duckduckgo-search>=8.1.1",
    "isort>=7.0.0",
    "langchain>=1.1.3",
    "langchain-classic>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.1.3",
    "langchain-openai>=1.1.1",
    "python-dotenv>=1.2.1",
]
